+++ src/AudioLib/Generator.h
@@ -6,7 +6,13 @@ class Generator
 {
 public:
 	virtual ~Generator() = default;
-	virtual int generate(int16_t *outBuffer) = 0;
+	/**
+	 * @brief Generates samples and copies them to the supplied buffer
+	 * 
+	 * @param outBuffer Specified buffer for the generated samples.
+	 * @return size_t Number of samples generated.
+	 */
+	virtual size_t generate(int16_t *outBuffer) = 0;
 	virtual int available() = 0;
 };
