+++ src/AudioLib/SourceWAV.h
@@ -11,7 +11,7 @@ public:
 	SourceWAV();
 	SourceWAV(fs::File *file);
 	~SourceWAV();
-	int generate(int16_t* outBuffer) override;
+	size_t generate(int16_t* outBuffer) override;
 	int available() override;
 
 	int getBitsPerSample() override;
